+++ Cargo.toml
@@ -10,3 +10,4 @@ regex = "1"
 glob = "0.3.0"
 rayon = "1.5.1"
 thiserror = "1.0.52"
+jwalk = "0.5"
