+++ crates/good_fences/Cargo.toml
@@ -12,7 +12,7 @@ name = "unused"
 path = "src/bin/unused.rs"
 
 [dependencies]
-jwalk = "0.5"
+jwalk.workspace = true
 lazy_static = "1.4.0"
 serde_derive = "1.0.117"
 serde_bytes = "0.11.5"
