+++ crates/unused_finder/src/walked_file.rs
@@ -1,6 +1,7 @@
+use crate::import_export_info::ImportExportInfo;
 
 #[derive(Debug, PartialEq, Eq, Clone)]
-pub struct WalkFileMetaData {
+pub struct WalkedSourceFile {
     pub package_name: String,
     pub source_file_path: String,
     pub import_export_info: ImportExportInfo,
@@ -8,7 +9,7 @@ pub struct WalkFileMetaData {
 
 #[derive(Debug, PartialEq)]
 pub enum WalkedFile {
-    SourceFile(WalkFileMetaData),
+    SourceFile(WalkedSourceFile),
     Nothing,
 }
